Begin apply database change

+++ src/java/com/development/simplestockmanager/business/persistence/controller/BrandJpaController.java
@@ -6,15 +6,18 @@
 package com.development.simplestockmanager.business.persistence.controller;
 
 import com.development.simplestockmanager.business.persistence.Brand;
-import com.development.simplestockmanager.business.persistence.controller.exceptions.NonexistentEntityException;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import com.development.simplestockmanager.business.persistence.Product;
+import com.development.simplestockmanager.business.persistence.controller.exceptions.IllegalOrphanException;
+import com.development.simplestockmanager.business.persistence.controller.exceptions.NonexistentEntityException;
+import java.util.ArrayList;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
 
 /**
  *
@@ -32,11 +35,29 @@ public class BrandJpaController implements Serializable {
     }
 
     public void create(Brand brand) {
+        if (brand.getProductList() == null) {
+            brand.setProductList(new ArrayList<Product>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Product> attachedProductList = new ArrayList<Product>();
+            for (Product productListProductToAttach : brand.getProductList()) {
+                productListProductToAttach = em.getReference(productListProductToAttach.getClass(), productListProductToAttach.getId());
+                attachedProductList.add(productListProductToAttach);
+            }
+            brand.setProductList(attachedProductList);
             em.persist(brand);
+            for (Product productListProduct : brand.getProductList()) {
+                Brand oldBrandOfProductListProduct = productListProduct.getBrand();
+                productListProduct.setBrand(brand);
+                productListProduct = em.merge(productListProduct);
+                if (oldBrandOfProductListProduct != null) {
+                    oldBrandOfProductListProduct.getProductList().remove(productListProduct);
+                    oldBrandOfProductListProduct = em.merge(oldBrandOfProductListProduct);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -45,12 +66,45 @@ public class BrandJpaController implements Serializable {
         }
     }
 
-    public void edit(Brand brand) throws NonexistentEntityException, Exception {
+    public void edit(Brand brand) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Brand persistentBrand = em.find(Brand.class, brand.getId());
+            List<Product> productListOld = persistentBrand.getProductList();
+            List<Product> productListNew = brand.getProductList();
+            List<String> illegalOrphanMessages = null;
+            for (Product productListOldProduct : productListOld) {
+                if (!productListNew.contains(productListOldProduct)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Product " + productListOldProduct + " since its brand field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Product> attachedProductListNew = new ArrayList<Product>();
+            for (Product productListNewProductToAttach : productListNew) {
+                productListNewProductToAttach = em.getReference(productListNewProductToAttach.getClass(), productListNewProductToAttach.getId());
+                attachedProductListNew.add(productListNewProductToAttach);
+            }
+            productListNew = attachedProductListNew;
+            brand.setProductList(productListNew);
             brand = em.merge(brand);
+            for (Product productListNewProduct : productListNew) {
+                if (!productListOld.contains(productListNewProduct)) {
+                    Brand oldBrandOfProductListNewProduct = productListNewProduct.getBrand();
+                    productListNewProduct.setBrand(brand);
+                    productListNewProduct = em.merge(productListNewProduct);
+                    if (oldBrandOfProductListNewProduct != null && !oldBrandOfProductListNewProduct.equals(brand)) {
+                        oldBrandOfProductListNewProduct.getProductList().remove(productListNewProduct);
+                        oldBrandOfProductListNewProduct = em.merge(oldBrandOfProductListNewProduct);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
@@ -68,7 +122,7 @@ public class BrandJpaController implements Serializable {
         }
     }
 
-    public void destroy(Long id) throws NonexistentEntityException {
+    public void destroy(Long id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -79,6 +133,17 @@ public class BrandJpaController implements Serializable {
                 brand.getId();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The brand with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Product> productListOrphanCheck = brand.getProductList();
+            for (Product productListOrphanCheckProduct : productListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Brand (" + brand + ") cannot be destroyed since the Product " + productListOrphanCheckProduct + " in its productList field has a non-nullable brand field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(brand);
             em.getTransaction().commit();
